Move `TestStore` into its own module
Now it is accessable as `crate::test::Store`. This allows to use it
outside the `collection` module where it used to be.
